#54 Count TextSpans using SQL not Java.
- Add three new NamedQueries to 'TextSpan'
- change the return type of getNumberOfTextSpansInStep() from int to
  long in super class 'OccurrencesService'
- Add methods performing the new queries to count the matching TextSpans
  to service classes 'EntityOccurrencesService',
  'AttributeOccurrencesService' and 'RelationOccurrencesService'

+++ src/main/java/de/unistuttgart/vis/vita/services/FingerprintsService.java
@@ -47,7 +47,7 @@ public class FingerprintsService extends OccurrencesService {
   }
 
   @Override
-  protected int getNumberOfSpansInStep(int stepStart, int stepEnd) {
+  protected long getNumberOfSpansInStep(int stepStart, int stepEnd) {
     // TODO not implemented yet!
     return 0;
   }
